tests: Fix sign compare warning in ASSERT_STR_EQ

+++ tests/knd-shard-test.c
@@ -29,10 +29,10 @@ static const char *shard_config =
         const char *__act = (act);                  \
         size_t __act_size = (act_size);             \
         const char *__exp = (exp);                  \
-        size_t __exp_size = ((exp_size) == (size_t)-1 ? strlen(__exp) : (size_t)(exp_size));     \
-        ck_assert_msg(__act_size == __exp_size && 0 == strncmp(__act, __exp, __act_size),        \
-            "Assertion '%s' failed: %s == \"%.*s\" [len: %zu] but expected \"%.*s\" [len: %zu]", \
-            #act" == "#exp, #act, __act_size, __act, __act_size, __exp_size, __exp, __exp_size); \
+        size_t __exp_size = ((size_t)(exp_size) == (size_t)-1 ? strlen(__exp) : (size_t)(exp_size)); \
+        ck_assert_msg(__act_size == __exp_size && 0 == strncmp(__act, __exp, __act_size),            \
+            "Assertion '%s' failed: %s == \"%.*s\" [len: %zu] but expected \"%.*s\" [len: %zu]",     \
+            #act" == "#exp, #act, __act_size, __act, __act_size, __exp_size, __exp, __exp_size);     \
     } while (0)
 
 struct table_test {
